update batch program
