+++ crates/types/base/Cargo.toml
@@ -11,7 +11,7 @@ version.workspace = true
 alloy-primitives = { workspace = true }
 alloy-serde.workspace = true
 rkyv.workspace = true
-sbv-primitives = { workspace = true, features = ["scroll-consensus-types", "scroll-reth-types", "scroll-reth-primitives-types", "reth-primitives-types"] }
+#sbv-primitives = { workspace = true, features = ["scroll-consensus-types", "scroll-reth-types", "scroll-reth-primitives-types", "reth-primitives-types"] }
 serde.workspace = true
 itertools.workspace = true
 sha3 = "0.10.8"
