+++ crates/types/base/src/utils/hash.rs
@@ -1,7 +1,5 @@
 use alloy_primitives::B256;
 
-/// From the utility of ether-rs
-///
 /// Computes the Keccak-256 hash of input bytes.
 ///
 /// Note that strings are interpreted as UTF-8 bytes,
