+++ crates/types/batch/src/blob_consistency/mod.rs
@@ -2,10 +2,11 @@ use alloy_primitives::{B256 as H256, U256};
 
 use openvm_pairing::bls12_381::Scalar;
 
+mod constants;
 mod openvm;
 pub use openvm::point_evaluation;
 
-pub use openvm::{EccToPairing, kzg_to_versioned_hash, verify_kzg_proof};
+pub use openvm::{ToIntrinsic, kzg_to_versioned_hash, verify_kzg_proof};
 
 // Number of bytes in a u256.
 pub const N_BYTES_U256: usize = 32;
